dishes fetch removed to layout
