+++ resources/js/Routes.js
@@ -1,17 +1,15 @@
 import React from "react";
 import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
-import Header from "./components/header";
-import { DishCardList } from "./components/dishcardlist";
-import { Dish } from "./components/dish";
+import Layout from "./components/layout";
 
 const routes = [
    {
       path: "/",
-      component: DishCardList,
+      component: Layout,
    },
    {
-      path: "/dish",
-      component: Dish,
+      path: "/dish/:dishId",
+      component: Layout,
 
    }
 ];
@@ -20,7 +18,6 @@ export const Routes = () => {
    return (
       <Router>
          <div>
-            <Header />
             <Switch>
                { routes.map((route) => (
                   <Route key={ route.path } path={ route.path } component={ route.component } exact />
