+++ resources/js/components/dishcardlist/index.js
@@ -1,20 +1,9 @@
-import React, { useEffect, useState } from 'react';
+import React from 'react';
 import "./styles.css"
 import { DishCard } from './dishcard';
 
 
 export const DishCardList = () => {
-
-   const [dishes, setDishes] = useState([])
-
-   const getDishes = () =>
-      window.axios.get('/full-dishes')
-         .then(json => setDishes(json.data.data))
-
-
-   useEffect(() => {
-      getDishes()
-   }, [])   
 
    return (
       <ul className="list">
