+++ resources/js/components/layout/index.js
@@ -1,15 +1,30 @@
 import React from 'react';
 import './styles.css'
 import Header from '../header'
+import { DishCardList } from '../dishcardlist';
+import { Dish } from '../dish';
 
 const Layout = () => {
-    return (<>
-        <Header/>
-        <main className='layout-content'>
-            <p>Lorem ipsum dolor sit amet, consectetur adipisicing elit. Eligendi, nisi?</p>
-            <p>Lorem ipsum dolor sit amet, consectetur adipisicing elit. Eaque id minima molestiae nisi, quo quos.</p>
-        </main>
-    </>)
+
+    const [dishes, setDishes] = useState([])
+
+    const getDishes = () =>
+        window.axios.get('/full-dishes')
+            .then(json => setDishes(json.data.data))
+
+    useEffect(() => {
+        getDishes()
+    }, [])
+
+    const dishId = useParams()
+
+    return (
+        <>
+            <Header />
+            <main className='layout-content'>
+                { dishId == undefined ? <DishCardList  /> : <Dish /> }
+            </main>
+        </>)
 }
 
 export default Layout
